Update create_trajectories_kml.R
This line caused errors for me in the later "for (tt in tidxs)" loops. The timestamps get converted to numeric in the loop, causing it to throw an "out of bonds" error. This is a simple proposed change that fixed it for me. This actually gets the indeces instead of the timestamps.